adds test for displaying todos in the dom
