+++ package.json
@@ -27,6 +27,7 @@
     "eslint-config-airbnb-base": "^14.2.1",
     "eslint-plugin-import": "^2.24.0",
     "html-webpack-plugin": "^5.3.2",
+    "identity-obj-proxy": "^3.0.0",
     "jest": "^27.0.6",
     "style-loader": "^3.2.1",
     "stylelint": "^13.13.1",
@@ -39,6 +40,7 @@
   },
   "dependencies": {
     "@jest/globals": "^27.0.6",
+    "jsdom": "^16.7.0",
     "lodash": "^4.17.21"
   }
 }
