+++ src/index.html
@@ -8,7 +8,7 @@
 </head>
 
 <body>
-    <section class="body">
+    <section class="myBody">
         <div id="todos">
             <h3>Today's To Dos</h3>
             <form>
