+++ src/style.css
@@ -1,5 +1,5 @@
 
-.body {
+.myBody {
   display: flex;
   justify-content: center;
 }
